replacing method from controller to service
