+++ core/src/main/java/greencity/controller/OrderController.java
@@ -6,9 +6,7 @@ import greencity.configuration.RedirectionConfigProp;
 import greencity.constants.HttpStatuses;
 import greencity.constants.ValidationConstant;
 import greencity.dto.*;
-import greencity.entity.order.Order;
 import greencity.entity.user.User;
-import greencity.repository.OrderRepository;
 import greencity.service.ubs.NotificationService;
 import greencity.service.ubs.UBSClientService;
 import io.swagger.annotations.ApiOperation;
@@ -27,7 +25,6 @@ import javax.validation.constraints.Pattern;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("/ubs")
@@ -36,19 +33,16 @@ public class OrderController {
     private final UBSClientService ubsClientService;
     private final RedirectionConfigProp redirectionConfigProp;
     private final NotificationService notificationService;
-    private final OrderRepository orderRepository;
 
     /**
      * Constructor with parameters.
      */
     @Autowired
     public OrderController(UBSClientService ubsClientService, RedirectionConfigProp redirectionConfigProp,
-        NotificationService notificationService,
-        OrderRepository orderRepository) {
+        NotificationService notificationService) {
         this.ubsClientService = ubsClientService;
         this.redirectionConfigProp = redirectionConfigProp;
         this.notificationService = notificationService;
-        this.orderRepository = orderRepository;
     }
 
     /**
@@ -152,7 +146,7 @@ public class OrderController {
         PaymentResponseDto dto, HttpServletResponse response) throws IOException {
         ubsClientService.validatePayment(dto);
         if (HttpStatus.OK.is2xxSuccessful()) {
-            notifyPaidOrder(dto);
+            notificationService.notifyPaidOrder(dto);
             response.sendRedirect(redirectionConfigProp.getGreenCityClient());
         }
         return ResponseEntity.status(HttpStatus.OK).build();
@@ -458,19 +452,9 @@ public class OrderController {
         PaymentResponseDto dto, HttpServletResponse response) throws IOException {
         ubsClientService.validatePaymentClient(dto);
         if (HttpStatus.OK.is2xxSuccessful()) {
-            notifyPaidOrder(dto);
+            notificationService.notifyPaidOrder(dto);
             response.sendRedirect(redirectionConfigProp.getGreenCityClient());
         }
         return ResponseEntity.status(HttpStatus.OK).build();
-    }
-
-    private void notifyPaidOrder(PaymentResponseDto dto) {
-        if (dto.getOrder_id() != null) {
-            Long orderId = Long.valueOf(dto.getOrder_id().split("_")[0]);
-            Optional<Order> orderOptional = orderRepository.findById(orderId);
-            if (orderOptional.isPresent()) {
-                notificationService.notifyPaidOrder(orderOptional.get());
-            }
-        }
     }
 }
